pyproject.toml
[tool.poetry]
name = "projeto-webscrap-fms"
version = "0.1.0"
description = ""
authors = ["FernandoMS4 <fernando381046@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
scrapy = "^2.12.0"
pandas = "^2.2.3"
sqlmodel = "^0.0.22"
mysql-connector-python = "^9.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
